pyproject.toml
[build-system]
requires = ["hatchling>=1.26", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
tag-pattern = "v*"

[project]
name = "easy-ha-satellite"
dynamic = ["version"]
description = "A plug-and-play satellite for Home Assistant using Docker"
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{ name = "Dave Bennett", email = "you@example.com" }]
license = { text = "MIT" }
urls = { "Source" = "https://github.com/davebben/easy-ha-satellite" }
classifiers = [
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
]

dependencies = [
  "sounddevice~=0.5.2",
  "openwakeword~=0.6.0",
  "numpy~=1.26.4",
  "websockets~=15.0.1",
  "uvloop~=0.21.0",
  "aiohttp~=3.12.14",
  "webrtc-noise-gain>=1.2.5",
  "soundfile~=0.13.1",
  "pydantic>=2.11.7",
  "pyaml>=25.7.0",
  "pydub>=0.25.1",
]

[project.optional-dependencies]
media = ["pydub~=0.25.1", "soundfile~=0.13.1"]

[project.scripts]
easy-ha-sat = "easy_ha_satellite.__main__:main"

[dependency-groups]
dev = [
  "ruff==0.12.5",
  "pytest>=8",
  "nox>=2024.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/easy_ha_satellite"]
include = ["src/easy_ha_satellite/assets/**/*"]
only-packages = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E4","E7","E9","F","I","UP","C4","B","SIM","PLE","PERF","NPY"]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
